requirements.txt
streamlit
torch
soundfile
nemo_toolkit[all]
